% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Confusion matrix.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Confusion matrix}
\usage{
confusion_matrix(X, y, bhat, cut = 0.5)
}
\arguments{
\item{y}{a numerical vector}

\item{bhat}{the estimated coefficients of the logistic regression}

\item{cut}{Default is 0.5}

\item{x}{a matrix of numerical coefficients of beta}
}
\value{
Confusion matrix
}
\description{
Confusion matrix
}
