% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output metrics.R
\name{Diagnostic_Odds_ration}
\alias{Diagnostic_Odds_ration}
\title{Diagnostic_Odds_ration}
\usage{
Diagnostic_Odds_ration(X, y, bhat, cut = 0.5)
}
\arguments{
\item{y}{a numerical vector}

\item{bhat}{the estimated coefficients of the logistic regression}

\item{cut}{Default is 0.5}

\item{x}{a matrix of numerical coefficients of beta}
}
\value{
Diagnostic_Odds_ration
}
\description{
Diagnostic_Odds_ration
}
