% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Betaestimate.R
\name{estBeta}
\alias{estBeta}
\title{Beta estimation
Calculate estimated beta}
\usage{
estBeta(X, y)
}
\arguments{
\item{X}{a matrix of beta}

\item{y}{a numerical vector}
}
\value{
list of beta and beta n
}
\description{
Beta estimation
Calculate estimated beta
}
