% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output metrics.R
\name{False_Discovery_Rate}
\alias{False_Discovery_Rate}
\title{False_Discovery_Rate}
\usage{
False_Discovery_Rate(X, y, bhat, cut = 0.5)
}
\arguments{
\item{y}{a numerical vector}

\item{bhat}{the estimated coefficients of the logistic regression}

\item{cut}{Default is 0.5}

\item{x}{a matrix of numerical coefficients of beta}
}
\value{
False_Discovery_Rate
}
\description{
False_Discovery_Rate
}
