% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intiabeta.R
\name{initialbeta}
\alias{initialbeta}
\title{Initial Beta}
\usage{
initialbeta(X, y)
}
\arguments{
\item{X}{a matrix of beta}

\item{y}{a numerical vector}
}
\value{
Initial Beta
}
\description{
Calculate Initial Beta
}
