% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap confidence interval.R
\name{interval}
\alias{interval}
\title{Bootstrap Confidence Interval}
\usage{
interval(X, y, alpha, B = 20)
}
\arguments{
\item{X}{a numeric dataset}

\item{y}{the significance value}

\item{n}{the number of bootstraps (default is 20)}
}
\value{
a matrix containing the bootstrap confidence interval
}
\description{
Bootstrap Confidence Interval
}
