% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logisticcurve.R
\name{logiplot}
\alias{logiplot}
\title{Logistic curve}
\usage{
logiplot(X, y, beta)
}
\arguments{
\item{X}{a matrix of numerical coefficient of beta}

\item{y}{a numerical vector}

\item{beta}{the coeffcient vector}
}
\description{
This function is used for Ploting the logistic curve
}
