% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear Regression Model.R
\name{my_lm}
\alias{my_lm}
\title{"Linear Regression Model"}
\usage{
my_lm(response, predictors, alpha)
}
\arguments{
\item{response}{A vector in n*1 dimension denoting the true observed values}

\item{predictors}{A matrix in n*p dimension denoting the input values}

\item{alpha}{The significance level}

\item{beta}{The estimation of the coefficients}

\item{sigma2}{The estimation of the error}

\item{variance_beta}{The variance of estimated beta based on the input significance
level}

\item{ci}{The confidence interval of the estimated beta based on the input significance
leve}

\item{R-squared}{The coefficient of determination (including goodness of fit)}
}
\description{
Contains a linear regression function that also evaluates parameters such as beta hat, confidence 
intervals, and plots
}
