% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimfunction.R
\name{optimizing}
\alias{optimizing}
\title{Optimizing function}
\usage{
optimizing(X, y, beta)
}
\arguments{
\item{X}{a matrix of beta}

\item{y}{a numerical vector}

\item{beta}{the coefficient vector}
}
\value{
Sum of beta hat
}
\description{
Based on the equation to get beta
}
