% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output metrics.R
\name{Prevalence}
\alias{Prevalence}
\title{Prevalence}
\usage{
Prevalence(X, y, bhat, cut = 0.5)
}
\arguments{
\item{y}{a numerical vector}

\item{bhat}{the estimated coefficients of the logistic regression}

\item{cut}{Default is 0.5}

\item{x}{a matrix of numerical coefficients of beta}
}
\value{
Prevalence
}
\description{
Prevalence
}
