% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prevalence grid.R
\name{prevalencegrid}
\alias{prevalencegrid}
\title{plot Prevalence over a grid of cut-off value.}
\usage{
prevalencegrid(X, y, bhat, cut = 0.5)
}
\arguments{
\item{y}{a numerical vector}

\item{bhat}{the estimated coefficients of the logistic regression}

\item{cut}{Default is 0.5}

\item{x}{a matrix of numerical coefficients of beta}
}
\value{
table of prevalence
}
\description{
plot Prevalence over a grid of cut-off value.
}
