% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear Regression Model.R
\name{my_lm}
\alias{my_lm}
\title{"Linear Regression Model"}
\arguments{
\item{response}{Also known as beta, an estimation of the coefficients in the model}

\item{predictors}{Also known as sigma2, an estimation of the error}
}
\description{
Contains a linear regression function that also evaluates parameters such as beta hat, confidence 
intervals, and plots
}
